// Project_Apex/frontend/src/api/models/LongTermRoiPushResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LongTermRoiPushResponse = {
    success: boolean;
    message: string;
    affected_users: number;
    total_roi_amount: number;
    execution_event_id: string;
};